requirements.txt
# ==========================================
# Core Framework
# ==========================================
fastapi==0.115.0
uvicorn[standard]==0.32.0

# ==========================================
# LangChain & LLM
# ==========================================
langchain==0.3.7
langchain-core==0.3.15
langchain-groq==0.2.1
groq==0.11.0

# ==========================================
# Git Operations
# ==========================================
GitPython==3.1.43

# ==========================================
# HTTP Requests (for GitHub API)
# ==========================================
requests==2.32.3

# ==========================================
# YAML Processing
# ==========================================
PyYAML==6.0.2

# ==========================================
# Data Validation
# ==========================================
pydantic==2.9.2

# ==========================================
# Environment Variables
# ==========================================
python-dotenv==1.0.1

# ==========================================
# CORS Support
# ==========================================
# (included in fastapi, but explicit for clarity)
# python-multipart==0.0.9  # if you need file uploads later